+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Button to copy recycle bin name string to clipboard in the import KeePassX 2 section
 - Confirmation dialog for remove Premium Key action
 - Confirmation dialog for discard item and item authorization changes action
+- Password generator dialog on item screen
 
 ## Changed
 - Allow copying item details of read-only items (do not disable them anymore) - the save and delete button is hidden instead of disabled to avoid user confusion